Adding Jobs.c and Jobs.h which will handle a dynamic list of jobs
These jobs are going to be used for background processes so that the
job id, process id (pid) and command can be printed to the screen.

This commit is to keep record of all of te changes that were made.
I have purposefully left in a lot of comments so that I can keep a
record of them.